BACKENDS: ATARI: Separate AtariEventSource from OSystem_Atari
